+++ src/main/java/com/example/demo_shop/model/Product.java
@@ -15,9 +15,9 @@ import java.math.BigDecimal;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     private String description;
     private BigDecimal price;
-    private int stock;
+    private long stock;
 }
